remove interface from interfaces folder and put in components
