--- src/interfaces/defaultLayout.ts
@@ -1,5 +1,0 @@
-import { ReactNode } from 'react'
-
-export interface ChildrenComponent {
-  children: ReactNode
-}
