--- src/interfaces/headerSkeletonPage.ts
@@ -1,6 +1,0 @@
-export interface HeaderSkeletonPageProps {
-  header: string
-  description: string
-  buttonText?: string
-  buttonClickHandler?: React.MouseEventHandler<HTMLButtonElement>
-}
