--- src/interfaces/menuItem.ts
@@ -1,5 +1,0 @@
-export interface MenuItemProps {
-  title: string
-  Icon: React.ElementType
-  link: string
-}
