--- src/interfaces/menuItems.ts
@@ -1,5 +1,0 @@
-export interface MenuItem {
-  title: string
-  icon: React.ElementType
-  link: string
-}
