--- src/interfaces/searchInput.ts
@@ -1,5 +1,0 @@
-export interface SearchInputProps {
-  value: string
-  onChange: (event: React.ChangeEvent<HTMLInputElement>) => void
-  placeholder: string
-}
