--- src/interfaces/skeletonPage.ts
@@ -1,7 +1,0 @@
-export interface SkeletonPageProps {
-  header: string
-  description: string
-  buttonText?: string
-  buttonClickHandler?: React.MouseEventHandler<HTMLButtonElement>
-  table: React.ReactNode
-}
